hide the js running console
When compiling a GUI app using pyinstaller with -w (or --noconsole or --windowed) and run it, I can see a console window for a millisecond and then it shutdown. The GUI app is run with no error.
The app use pyexecjs for running js.
I need to hide the console window  when js is runing.